add control to force non-profiling command queues
This can be used for performance analysis, but may cause errors if
an application requires event profiling.  If this control is set
along with another control that requires event profiling, such as
DevicePerformanceTiming, then the other control will override this
control, but this behavior may change in the future and should not
be relied upon!