adding ChromePerformanceTimingPerKernel control
ChromePerformanceTimingPerKernel: This is a further control that sits
on top of ChromePerformanceTiming (it does nothing when this flag is not
on). When it is on, it organizes the performance information placed in the
JSON file on a per kernel name basis. It can be combined with the
ChromePerformanceTimingInStages control for information about event stages.

This control can be enabled when one timeline row per kernel name
is desired in chrome traces.