add basic support for SVM and USM capture and replay (#419)
* add basic support for SVM and USM capture and replay

There are still limitations, specifically when setting a pointer
to the middle of an SVM or USM allocation as a kernel argument, or
when accessing SVM or USM indirectly.  Basic cases should be working,
though, that set a pointer to the base of an SVM or USM allocation as
a kernel argument.

* record the SVM and USM pointer rather than the data it points to